misc update

+++ CSV2SerialAdapter/src/gde/device/wb/CSV2SerialAdapter.java
@@ -296,10 +296,10 @@ public class CSV2SerialAdapter extends DeviceConfiguration implements IDevice {
 
 		try {
 			if (subtractFirst) {
-				reduction = record.getFirst().intValue() / 1000.0;
+				reduction = record.getFirst() / 1000.0;
 			}
 			else if (subtractLast) {
-				reduction = record.getLast().intValue() / 1000.0;
+				reduction = record.getLast() / 1000.0;
 			}
 		}
 		catch (Throwable e) {
@@ -328,10 +328,10 @@ public class CSV2SerialAdapter extends DeviceConfiguration implements IDevice {
 
 		try {
 			if (subtractFirst) {
-				reduction = record.getFirst().intValue() / 1000.0;
+				reduction = record.getFirst() / 1000.0;
 			}
 			else if (subtractLast) {
-				reduction = record.getLast().intValue() / 1000.0;
+				reduction = record.getLast() / 1000.0;
 			}
 		}
 		catch (Throwable e) {
